requirements.txt
# Map Matcher UI Dependencies
PySide6
PyYAML
opencv-python

# VAID GIS (local package)
# Install with: pip install -e packages/vaid_gis
# Dependencies for vaid_gis:
numpy
geopandas
shapely
scipy

# Deep Learning / Segmentation
# Use --index-url https://download.pytorch.org/whl/cu118 (or appropriate CUDA version) if needed manually
torch
torchvision
transformers
pillow
